fix(UserService):handle non existing user - deleteUserById()

+++ src/main/java/spring/project/adi/service/UserService.java
@@ -37,6 +37,8 @@ public class UserService {
     }
 
     public void deleteUserById(UUID userId) {
-        this.userRepository.deleteById(userId);
+        if (this.userRepository.existsById(userId)) {
+            this.userRepository.deleteById(userId);
+        }
     }
 }
